Fix: missed a float conversion

+++ src/rust/lqos_python/src/device_weights.rs
@@ -77,7 +77,7 @@ fn get_weights_from_shaped_devices() -> Result<Vec<DeviceWeightResponse>> {
             continue;
         }
         // Convert f32 to weight with proper rounding and minimum safeguards
-        let weight = f32::max(1.0, device.download_max_mbps / 2.0).round() as i64;
+        let weight = f32::max(1.0, device.download_max_mbps as f32 / 2.0).round() as i64;
         result.push(DeviceWeightResponse {
             circuit_id: circuit_id.clone(),
             weight,
